ipfixsend: fix global vars set at signal handler
If we change global variables asynchronously, out of the scope of a
function, the compiler needs to be aware of the fact, otherwise an
optimization will leave out checks on these variables, which usually
take place in the loops.  The reason is that the compiler sees no change
to the variable inside the loop, thus it can optimize the check of a
variable that doesn't change.  `volatile' modifier signals to the
compiler that this variable might change from another scope than the one
the compiler sees, thus it forces the compiler to re-fetch the
variable's value every time the check is performed.

This is a common case with flags in signal handlers, as the signal is an
asynchronous way of changing the status of a program.