queues: fix Doxygen warning about unused param
A really strange warning coming out of Doxygen, which complains about a
wrong parameter name in documentation.  Yet, after checking the
definition of the function, the parameter's name was properly defined.

The warning message from Doxygen was the following:

   base/src/queues.c:119: warning: argument 'refcount' of command @param
      is not found in the argument list of
      rbuffer_write(struct ring_buffer *rbuffer, struct ipfix_message *record, uint16_t ref_count)

The problem stems from a deeper and more important issue:  duplication
of documentation between headers and compilation units.  If there are
duplicated doc-comments, Doxygen will pick up one, if they are
identical, otherwise it will merge both of them.  So, normally there is
no problem for Doxygen, but there is a huge maintenance cost for the
developers.  In this particular case, the developer renamed the
function's parameter *and* updated the doc-comment in the .c file, as
both were rather close.  Yet, he didn't realize that there was yet
another definition of the same comment, which was far away from the
point of the refactoring.

In short, this patch solves only the specific problem, but we should
address the whole problem of duplicated doc-comments.  Only one version
should exist in the codebase, ideally the one that is close to the
function/type definition.  After all, header files are meant to be
consumed by compilers, not by humans.  For this reason, we have
Doxygen's documentation!  On the other hand, it's far more possible for
a developer to update a comment, if it is close enough to the change in
the code.

Alternatively, we could update the doc-comment in the header file to
look the same as in the .c, but this would solve the problem partially
and temporarily.