sflow.c: clear out buffer without stack allocation
A puzzled snippet of code.

In readExtendedSocket4(), the `buf` is only declared *and*
zero-initialized with the only intention to clear out the
`sfCLF.client`.  We can certainly do better!  At least, we avoid growing
unnecessary the stack, with the risk of overflowing it.

By the way, notice also the inconsistency in clearing out the last
element of the buffer!  Yet, both of them were, technically, correct.