fbitmerge: flush output stream only once
I/O operations are always slow, and in this case, we abandon the I/O
buffering, as we call the `std::endl` after each write.  Since we are
going to write, essentially, 3 lines of small text, which can easily fit
in an I/O buffer, and then close the stream, it's better to drop
cache-flushing altogether and perform it once at the end.