fbitmerge: drop stringstream, use string assignment
We don't need such a heavy and complex operation of using a
`stringstream`, only to copy a C string into a `std::string` variable.
The assignment operator of `std::string` can take care of this operation
just fine.

Actually, it would be even better had we initialized directly the
`std::string` instances with the required C strings, but this is
slightly difficult in a scope of a `switch`.